requirements.txt
Django==4.0.3
djangorestframework==3.12.4
pytz==2022.1